[BUGFIX] tests without sleep
