+++ database/all/internal.go
@@ -1,74 +1,41 @@
 package all
 
 import (
-	"log"
 	"time"
 
 	"github.com/FreifunkBremen/yanic/database"
 	"github.com/FreifunkBremen/yanic/runtime"
 )
 
-type Connection struct {
-	database.Connection
-	list []database.Connection
+var conn database.Connection
+var quit chan struct{}
+
+func Start(config runtime.DatabaseConfig) (err error) {
+	conn, err = Connect(config.Connection)
+	if err != nil {
+		return
+	}
+	quit = make(chan struct{})
+	go deleteWorker(config.DeleteInterval.Duration, config.DeleteAfter.Duration)
+	return
 }
 
-func Connect(allConnection map[string]interface{}) (database.Connection, error) {
-	var list []database.Connection
-	for dbType, conn := range database.Adapters {
-		configForType := allConnection[dbType]
-		if configForType == nil {
-			log.Printf("the output type '%s' has no configuration\n", dbType)
-			continue
+func Close() {
+	close(quit)
+	conn.Close()
+	quit = nil
+}
+
+// prunes node-specific data periodically
+func deleteWorker(deleteInterval time.Duration, deleteAfter time.Duration) {
+	ticker := time.NewTicker(deleteInterval)
+	for {
+		select {
+		case <-ticker.C:
+			conn.PruneNodes(deleteAfter)
+		case <-quit:
+			ticker.Stop()
+			return
 		}
-		dbConfigs, ok := configForType.([]map[string]interface{})
-		if !ok {
-			log.Panicf("the output type '%s' has the wrong format\n", dbType)
-		}
-
-		for _, config := range dbConfigs {
-			if c, ok := config["enable"].(bool); ok && !c {
-				continue
-			}
-			connected, err := conn(config)
-			if err != nil {
-				return nil, err
-			}
-			if connected == nil {
-				continue
-			}
-			list = append(list, connected)
-		}
-	}
-	return &Connection{list: list}, nil
-}
-
-func (conn *Connection) InsertNode(node *runtime.Node) {
-	for _, item := range conn.list {
-		item.InsertNode(node)
-	}
-}
-
-func (conn *Connection) InsertLink(link *runtime.Link, time time.Time) {
-	for _, item := range conn.list {
-		item.InsertLink(link, time)
-	}
-}
-
-func (conn *Connection) InsertGlobals(stats *runtime.GlobalStats, time time.Time, site string) {
-	for _, item := range conn.list {
-		item.InsertGlobals(stats, time, site)
-	}
-}
-
-func (conn *Connection) PruneNodes(deleteAfter time.Duration) {
-	for _, item := range conn.list {
-		item.PruneNodes(deleteAfter)
-	}
-}
-
-func (conn *Connection) Close() {
-	for _, item := range conn.list {
-		item.Close()
 	}
 }
